src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "pyodbc (>=5.2.0,<6.0.0)",
    "pytest (>=8.0.0,<9.0.0)"
]
[tool.poetry]
packages = [
    { include = "app.py" },
    { include = "iointerface.py" }
]
[tool.poetry.scripts]
neg_checker = "app:negative_position_checker"
dup_req = "app:dup_req"
short_strats = "app:short_strategies_checker"
contract_req = "app:cContractDurationReq"

[tool.pytest.ini_options]
pythonpath = [ "." ]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
